Use asyncio for connections instead of DBus
DBus is great as in IPC, but not as an RPC. You have to jump through a
lot of hoops to access a remote DBus bus. If this component is to be
used for RPC (e.g., in a container setup), then DBus isn't a great fit.
Furthermore, the pygi stack with Gio and GDBus isn't that common outside
of GNOME desktops.

Instead, use asyncio to handle connections over a TCP port. The RPC will
be redone in a subsequent commit.