Add message format and using msgpack
Define a simple messaging format for s3watcher. The message consists of
a header and body. The body is the message payload in MessagePack
format. The header begins with a single byte version number and then 4
bytes containing the payload size in big endian.

The format of the body is a tuple of (msgtype, parameters). msgtype is
integer enumeration of the message type. Parameters is an optional
dictionary. The keys and values in the dictionary are dependent on the
message type. The MessageType enum class holds the currently recognized
message types.

This also adds back the objects call returning a list of (key, size,
last modified) tuples.